fix(design-library): pin a stream to one file, and keep missing apart from forbidden
Two from the fifth review round, plus the file-size rule.

The size check could not see a file replaced by one of exactly the same
length: the file is reopened for every slice, so the bytes could come from two
different files and still satisfy every invariant. Each slice now carries the
file's identity — device, inode, size and mtime — and the read is abandoned if
it ever changes.

Sharing the containment helper had flattened two different answers into one:
a file that is simply missing was being reported as a path outside the storage,
which is both wrong and the more alarming of the two. The helper now says
which, and every caller passes it on.

Reading an item's files moves to its own module, with its own tests. The tool
surface was seven lines under the limit and its test file had gone over.
