+++ plugins/sero-design-library-plugin/extension/tools/tools.test.ts
@@ -357,128 +357,3 @@ describe('the asset tool will not follow a link out of its storage', () => {
     expect(textOf(result)).not.toContain('not yours');
   });
 });
-
-/**
- * Reading a stored clip in slices (D4).
- *
- * A still comes back whole and that is fine at thumbnail size. A clip is
- * megabytes, and the `data:` URL a whole-file read produces cannot be seeked or
- * streamed — which is a player that renders and will not play. So it is read in
- * pieces the renderer can assemble into a Blob.
- */
-describe('the asset tool streams an original in slices', () => {
-  const CONTENT = Buffer.alloc(UPLOAD_CHUNK_BYTES + 1024, 7);
-
-  async function seedClip(): Promise<string> {
-    const id = 'itm-clip';
-    const directory = path.join(paths.home, 'items', id);
-    await mkdir(directory, { recursive: true });
-    await writeFile(path.join(directory, 'original.mp4'), CONTENT);
-    await writeFile(
-      path.join(directory, 'record.json'),
-      JSON.stringify({
-        id,
-        schemaVersion: 1,
-        createdAt: 0,
-        updatedAt: 0,
-        kind: 'video',
-        source: { kind: 'generated', fileName: 'clip.mp4' },
-        asset: {
-          originalFile: 'original.mp4',
-          previewFile: 'poster.webp',
-          mediaType: 'video/mp4',
-          bytes: CONTENT.byteLength,
-          checksum: 'x',
-        },
-        analysis: { status: 'ready' },
-        collectionIds: [],
-      }),
-      'utf8',
-    );
-    return id;
-  }
-
-  it('hands back ordered pieces that rebuild the file exactly', async () => {
-    const id = await seedClip();
-    const parts: Buffer[] = [];
-    let offset = 0;
-    let total = 0;
-
-    for (let guard = 0; guard < 10; guard += 1) {
-      const result = await call('design_library_assets', { action: 'stream', itemId: id, offset });
-      const details = result.details as { total: number; bytes: number; data: string; mediaType: string };
-      expect(details.mediaType).toBe('video/mp4');
-      // Never more than one chunk at a time, whatever the file's size.
-      expect(details.bytes).toBeLessThanOrEqual(UPLOAD_CHUNK_BYTES);
-      total = details.total;
-      parts.push(Buffer.from(details.data, 'base64'));
-      offset += details.bytes;
-      if (offset >= details.total || details.bytes === 0) break;
-    }
-
-    expect(total).toBe(CONTENT.byteLength);
-    expect(Buffer.concat(parts)).toEqual(CONTENT);
-  });
-
-  it('says nothing is left rather than failing, past the end', async () => {
-    const id = await seedClip();
-    const result = await call('design_library_assets', {
-      action: 'stream',
-      itemId: id,
-      offset: CONTENT.byteLength + 5000,
-    });
-
-    expect(result.details).toMatchObject({ bytes: 0, total: CONTENT.byteLength });
-  });
-
-  it('refuses an offset that is not a byte position', async () => {
-    const id = await seedClip();
-    const result = await call('design_library_assets', { action: 'stream', itemId: id, offset: -1 });
-    expect(textOf(result)).toContain('not a valid offset');
-  });
-});
-
-describe('the asset tool will not follow a link out of its storage for an item', () => {
-  async function seedItemPointingAt(target: string): Promise<string> {
-    const id = 'itm-link';
-    const directory = path.join(paths.home, 'items', id);
-    await mkdir(directory, { recursive: true });
-    await symlink(target, path.join(directory, 'original.mp4'));
-    await writeFile(
-      path.join(directory, 'record.json'),
-      JSON.stringify({
-        id,
-        schemaVersion: 1,
-        createdAt: 0,
-        updatedAt: 0,
-        kind: 'video',
-        source: { kind: 'generated', fileName: 'clip.mp4' },
-        asset: {
-          originalFile: 'original.mp4',
-          previewFile: 'original.mp4',
-          mediaType: 'video/mp4',
-          bytes: 9,
-          checksum: 'x',
-        },
-        analysis: { status: 'ready' },
-        collectionIds: [],
-      }),
-      'utf8',
-    );
-    return id;
-  }
-
-  it('refuses to stream or read a file that links somewhere else', async () => {
-    const secret = path.join(home, 'outside.bin');
-    await writeFile(secret, 'not yours', 'utf8');
-    const id = await seedItemPointingAt(secret);
-
-    // Every reader of an item's files, not only the new one: the record names
-    // the file, and the record is not proof of where the file actually is.
-    for (const action of ['stream', 'original', 'preview']) {
-      const result = await call('design_library_assets', { action, itemId: id });
-      expect(textOf(result)).toContain('outside the Design Library directory');
-      expect(JSON.stringify(result)).not.toContain(Buffer.from('not yours').toString('base64'));
-    }
-  });
-});
